Add AVL Tree and Linked List Structures
This commit adds avl_tree.c and avl_tree.h files.

The avl_tree.h file contains the avl structure that we'll be using to track processes.
It also contains linked list structures that we will be using to track
processes and file descriptors.

The avl_tree.c file has functions that work with the standard avl_tree in
addition to functions that add file descriptors and children to each process and
a function to remove file descriptors from a process given its pid when they are
closed